refactor: using sourcemap as sourcemap builder
