+++ Cargo.toml
@@ -14,6 +14,7 @@ rustc-hash = { version = "1.1.0" }
 serde = { version = "1.0", features = ["derive"], optional = true }
 serde_json = { version = "1.0", optional = true }
 once_cell = "1.18.0"
+sourcemap = { version = "7.0.1" }
 
 [features]
 # Enable source map functionality
