+++ src/magic_string/source_map.rs
@@ -1,7 +1,7 @@
 use crate::{
     locator::Locator,
-    source_map::{decoded_map::DecodedMap, mappings::Mappings, SourceMap},
     MagicString,
+    sourcemap_builder::SourcemapBuilder,
 };
 
 #[derive(Debug, Default)]
@@ -10,60 +10,38 @@ pub struct SourceMapOptions {
 }
 
 impl<'s> MagicString<'s> {
-    fn generate_decoded_source_map(&self, opts: SourceMapOptions) -> DecodedMap {
-        let mut mappings = Mappings::new();
+    pub fn source_map(&self, opts: SourceMapOptions) -> sourcemap::SourceMap {
+        let mut source_builder = SourcemapBuilder::new();
+
+        source_builder.set_source("");
+        if opts.include_content {
+            source_builder.set_source_contents(&self.source);
+        }
+
         let locator = Locator::new(&self.source);
 
         self.intro.iter().for_each(|frag| {
-            mappings.advance(frag);
+            source_builder.advance(frag);
         });
-
-        let mut names = vec![];
 
         self.iter_chunks().for_each(|chunk| {
             chunk.intro.iter().for_each(|frag| {
-                mappings.advance(frag);
+                source_builder.advance(frag);
             });
 
-            let name_idx = if chunk.keep_in_mappings && chunk.is_edited() {
-                let original_content = chunk.span.text(&self.source);
-
-                let idx = names
-                    .iter()
-                    .enumerate()
-                    .find_map(|(idx, name)| (name == original_content).then_some(idx))
-                    .unwrap_or_else(|| {
-                        let next_idx = names.len();
-                        names.push(original_content.to_string());
-                        next_idx
-                    });
-                debug_assert!(idx < names.len());
-                Some(idx as u32)
+            let name = if chunk.keep_in_mappings && chunk.is_edited() {
+                Some(chunk.span.text(&self.source))
             } else {
                 None
             };
 
-            mappings.add_chunk(chunk, &locator, 0, &self.source, name_idx);
+            source_builder.add_chunk(chunk, &locator, &self.source, name);
 
             chunk.outro.iter().for_each(|frag| {
-                mappings.advance(frag);
+                source_builder.advance(frag);
             });
         });
 
-        DecodedMap {
-            version: 3,
-            sources: vec!["".to_string()],
-            sources_content: opts
-                .include_content
-                .then(|| vec![self.source.as_ref().to_string()])
-                .unwrap_or_default(),
-            mappings,
-            names,
-        }
-    }
-
-    pub fn source_map(&self, opts: SourceMapOptions) -> SourceMap {
-        let decoded_map = self.generate_decoded_source_map(opts);
-        decoded_map.into_source_map()
+        source_builder.into_source_map()
     }
 }
